Fix `etcd.status.clusterSize` update (#260)
* Use int32 for spec.replicas

Using `int` is not encouraged by Kubernets API conventions.
https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#primitive-types.

* Maintain etcd.status.replicas

* Fix status.clusterSize update

+++ controllers/etcd_controller_test.go
@@ -232,7 +232,7 @@ var _ = Describe("Druid", func() {
 					return nil, err
 				}
 				return instance.Status.ClusterSize, nil
-			}, timeout, pollingInterval).Should(Equal(pointer.Int32Ptr(int32(instance.Spec.Replicas))))
+			}, timeout, pollingInterval).Should(Equal(pointer.Int32Ptr(instance.Spec.Replicas)))
 		})
 		It("should create and adopt statefulset and printing events", func() {
 			// Check StatefulSet requirements
